chore: refactor user operator files
