+++ client/app/(authenticated)/bceidbusiness/industry_user_admin/dashboard/users/page.tsx
@@ -1,11 +1,11 @@
 import { Suspense } from "react";
 
-import UserOperatorDataGrid from "@/app/components/datagrid/UserOperatorDataGrid";
 import Loading from "@/app/components/loading/SkeletonGrid";
 import {
   processExternalDashboardUsersTileData,
   UserOperatorDataGridRow,
 } from "@/app/utils/users/adminUserOperators";
+import UserOperatorDataGrid from "@/app/components/userOperators/UserOperatorDataGrid";
 
 const UserOperatorsPage = async () => {
   const userOperatorData: { rows: UserOperatorDataGridRow[] } =
