+++ client/app/(authenticated)/idir/cas_admin/dashboard/operators/user-operator/[id]/[formSection]/page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
 import Loading from "@/app/components/loading/SkeletonGrid";
-import UserOperator from "@/app/components/routes/select-operator/form/UserOperator";
+import UserOperator from "@/app/components/userOperators/UserOperator";
 
 export default async function Page({ params }: { params: { id: string } }) {
   return (
