+++ client/app/components/routes/select-operator/Page.tsx
@@ -1,7 +1,6 @@
 import { Suspense } from "react";
 import { permanentRedirect } from "next/navigation";
 import Loading from "@/app/components/loading/SkeletonField";
-import SelectOperator from "@/app/components/routes/select-operator/form/SelectOperator";
 import { BC_GOV_LINKS_COLOR } from "@/app/styles/colors";
 import Link from "next/link";
 import { getServerSession } from "next-auth";
@@ -9,13 +8,14 @@ import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { actionHandler } from "@/app/utils/actions";
 import { OperatorStatus, UserOperatorStatus } from "@/app/utils/enums";
 import getUserFullName from "@/app/utils/getUserFullName";
+import SelectOperator from "../../userOperators/SelectOperator";
 
 export const getUserOperator = async () => {
   try {
     return await actionHandler(
       `registration/user-operator/user-operator-from-user`,
       "GET",
-      "",
+      ""
     );
   } catch (error) {
     throw error;
@@ -33,7 +33,7 @@ export default async function MyOperatorPage() {
     // Using permanentRedirect instead of redirect to avoid the double rendering bug
     // https://github.com/vercel/next.js/issues/57257
     return permanentRedirect(
-      `/dashboard/select-operator/user-operator/${id}/1?title=${operatorLegalName}`,
+      `/dashboard/select-operator/user-operator/${id}/1?title=${operatorLegalName}`
     );
   }
 
@@ -43,11 +43,11 @@ export default async function MyOperatorPage() {
   ) {
     if (isNew) {
       return permanentRedirect(
-        `/dashboard/select-operator/received/add-operator/${operatorId}?title=${operatorLegalName}`,
+        `/dashboard/select-operator/received/add-operator/${operatorId}?title=${operatorLegalName}`
       );
     }
     return permanentRedirect(
-      `/dashboard/select-operator/received/request-access/${operatorId}?title=${operatorLegalName}`,
+      `/dashboard/select-operator/received/request-access/${operatorId}?title=${operatorLegalName}`
     );
   }
