+++ client/app/components/userOperators/SelectOperator.tsx
@@ -1,5 +1,5 @@
-import SelectOperatorForm from "@/app/components/form/SelectOperatorForm";
 import { selectOperatorSchema } from "@/app/utils/jsonSchema/selectOperator";
+import SelectOperatorForm from "./SelectOperatorForm";
 
 export default async function SelectOperator() {
   return <SelectOperatorForm schema={selectOperatorSchema} />;
