+++ client/app/components/userOperators/types.ts
@@ -1,3 +1,25 @@
+export interface OperatorsSearchParams {
+  [key: string]: string | number | undefined;
+  page?: number;
+  sort_field?: string;
+  sort_order?: string;
+}
+
+export interface UserOperator {
+  id: number;
+  status: string;
+  first_name: string;
+  last_name: string;
+  email: string;
+  legal_name: string;
+}
+
+export interface UserOperatorPaginated {
+  rows: UserOperator[];
+  row_count: number;
+  page?: number;
+}
+
 export interface Operator {
   id: number;
   legal_name: string;
